Count the numbers
Write a program that reads a, b, and n and outputs how many numbers in the inclusive range from a to b (a < b) are divisible by n.

Note: you can write this program in a more efficient way, that is, without any loops. Think about it after you write the solution with a loop.